+++ Source/Ariadne/CurrentlyVisibleViewFinder.swift
@@ -71,6 +71,57 @@ open class CurrentlyVisibleViewFinder: ViewFinder {
     }
 }
 
+let printSplitViewControllerWarning: Void = {
+    print("""
+    ⚠️ Attempting to search UISplitViewController that is collapsed. This leads to selection of only one controller, which `SplitViewCurrentlyVisibleViewFinder` will still return.
+    This, however, can lead to incorrect logic, since you might get a detail view controller when requesting master and vice versa.
+    This warning will be printed only once.
+""")
+}()
+
+/// `CurrentlyVisibleViewFinder` that proritizes either master or detail view controller in split view controllers.
+///  Please note that this ViewFinder should only be used for iPad interfaces that do not have collapsed UISplitViewControllers.
+///  If UISplitViewController is collapsed, it's `viewControllers` property only returns one of two(master,detail) view controllers and therefore it's impossible to know if one of those is a master or a detail.
+open class SplitViewCurrentlyVisibleViewFinder: CurrentlyVisibleViewFinder {
+
+    /// Kind of view controller that should be prioritized when searching for currently visible view controller
+    public enum Kind {
+        case master
+        case detail
+    }
+
+    /// Kind of view controller that should be prioritized when searching for currently visible view controller
+    public let kind: Kind
+
+    /// Creates `SplitViewCurrentlyVisibleViewFinder` that prioritizes `kind` view controller when searching for currently visible view controller.
+    /// - Parameters:
+    ///   - kind: master or detail view
+    ///   - rootViewProvider: provider of root view in hierarchy
+    public init(kind: Kind, rootViewProvider: RootViewProvider?) {
+        self.kind = kind
+        super.init(rootViewProvider: rootViewProvider)
+    }
+
+     /// Recursively searches view hierarchy for currently visible view. When `UISplitViewController` is encountered, `kind` is prioritized (.master: viewControllers.first, .detail: viewControllers.last)
+       ///
+       /// - Parameter view: view to start search from
+       /// - Returns: currently visible view.
+    open override func findCurrentlyVisibleView(startingFrom view: ViewController?) -> ViewController? {
+        guard let splitView = view as? UISplitViewController else {
+            return super.findCurrentlyVisibleView(startingFrom: view)
+        }
+        if splitView.viewControllers.count > 1 {
+            printSplitViewControllerWarning
+        }
+        switch kind {
+        case .master:
+            return splitView.viewControllers.first
+        case .detail:
+            return splitView.viewControllers.last
+        }
+    }
+}
+
 #endif
 
 #endif